mqtt/requirements.txt
paho-mqtt
pymongo
PyYAML
